Added updated sql plus prompt for deleting products

+++ classes/navigation.php
@@ -52,7 +52,7 @@ class Navigation extends Form{
 			echo '<p>'.substr($product['description'],0,150).'...';
 			echo '<span class="price">$'.$product['price'].'</span></p>';
 			echo '<a href="product_edit_page.php?id='.$product['product_id'].'" class="edit">Edit Page</a>';
-			echo '<a href="delete_page.php?id='.$product['product_id'].'" name="delete">Delete</a>';
+			echo '<a href="delete_page.php?id='.$product['product_id'].'" name="delete" onclick="return confirm(\'Are you sure you want to delete this?\');">Delete</a>';
 			echo '</a></div>';
 
 		}
